Initial extraction of token structures from parser
